requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
djangorestframework==3.14.0
idna==3.4
prettyconf==2.2.1
psycopg2-binary==2.9.7
pycparser==2.21
pyOpenSSL==23.2.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
